Added 'Poppable' Monobehavior and 'PopManager' singleton
Poppable is meant to sit on the top level structure of anything that can pop. (Right now it's tied a bit to the robot so we might modify when other poppables come)
Poppables listen for triggers which will kick off their own pop. Then they queue themselves up in the PopManager.

The PopManager consumes the queue and kicks off events in vfx graph for the poppable as well as initiating the actual pop event for the poppable

The pop function within Poppables will start the pop sequence animating textures, activating rigidbodies, setting objects inactive, and searching for other poppables to interact with and add to the queue.
These can be extended with other poppable patterns (ways to trigger other poppables)