Feature/nap bubble shader (#14)
* Enabling WebGPU in the build settings

- Enabled WebGPU
- Set default web player to WebGPU. Still falling back on WebGL, but this could cause Firefox/Safari users to have a subpar experience as they need to enable WebGPU via a developer flag (Would error when playing if we didn't fall back on WebGL)
- Set the default build settings for web to the PC settings (on the default (mobile settings) we don't have access to some rendering features we will need). We might need to tweak these settings later if they are too intensive for some players.

* Added Visual Effects Graph

* Ignoring Rider IDE files

* Adding Bubble Materials and Shaders

* Setting up scene for testing

Only committing this so others can mess with the materials I have set up.

* Some more bubbles and a distortion material

* small update to a bubble

* Added Temp Test Assets -z

Added Temp test Assets for shaders

* Updated Materials

Added some textures from Zane
Added basic pop vfx

* Adding DoTween

* Pop VFX

* Added 'Poppable' Monobehavior and 'PopManager' singleton

Poppable is meant to sit on the top level structure of anything that can pop. (Right now it's tied a bit to the robot so we might modify when other poppables come)
Poppables listen for triggers which will kick off their own pop. Then they queue themselves up in the PopManager.

The PopManager consumes the queue and kicks off events in vfx graph for the poppable as well as initiating the actual pop event for the poppable

The pop function within Poppables will start the pop sequence animating textures, activating rigidbodies, setting objects inactive, and searching for other poppables to interact with and add to the queue.
These can be extended with other poppable patterns (ways to trigger other poppables)

* helper script for testing

* Updating scenes.

Adding test scene for the chained popping
Adding a Bubble Robot prefab (subject to change with the final model)

* Adding to namespace

* Update .gitattributes

* New formatting... And adding GetPoppablesInRange as the method to modify when inheriting

* Misc file updates. Nothing important

* empty commit for merge

* Updating gitattributes to match develop

---------

Co-authored-by: Zanejjb <mavicreates@gmail.com>